requirements.txt
torch
numpy
opencv-python
Pillow
kornia
